+++ packages/compiler/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @openfn/compiler
 
+## 0.0.41
+
+### Patch Changes
+
+- Updated dependencies [2fde0ad]
+  - @openfn/logger@1.0.1
+
 ## 0.0.40
 
 ### Patch Changes