+++ packages/lightning-mock/CHANGELOG.md
@@ -4,6 +4,12 @@
 
 ### Patch Changes
 
+- Updated dependencies [2fde0ad]
+- Updated dependencies [2fde0ad]
+  - @openfn/logger@1.0.1
+  - @openfn/engine-multi@1.0.1
+  - @openfn/lexicon@1.0.0
+  - @openfn/runtime@1.0.1
 - Updated dependencies [4f5f1dd]
 - Updated dependencies [58e0d11]
   - @openfn/engine-multi@1.1.0