+++ packages/runtime/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @openfn/runtime
 
+## 1.1.1
+
+### Patch Changes
+
+- Updated dependencies [2fde0ad]
+  - @openfn/logger@1.0.1
 ## 1.1.0
 
 ### Minor Changes